fixed integer field

+++ airsafe/airsafeapp/forms.py
@@ -1,8 +1,8 @@
 from django import forms
 
 class inputForm(forms.Form):
-    time = forms.IntegerField()
-    volume = forms.IntegerField()
+    time = forms.FloatField()
+    volume = forms.FloatField()
 
     class Meta:
         fields = ['time', 'volume']
